Create hook for fliprs
