+++ frontend/src/components/fliprs/FliprApp.tsx
@@ -1,23 +1,9 @@
 import FliprGallery from "./FliprGallery";
-import {useEffect, useState} from "react";
-import {Flipr} from "../models/Flipr";
-import axios from "axios";
+import useFliprs from "../api/useFliprs";
 
 export default function FliprApp(){
 
-    const [fliprs, setFliprs] = useState<Flipr[]>([])
-
-    useEffect(() => {
-        getFliprs()
-    }, [])
-
-    function getFliprs(){
-        axios.get("/api/fliprs")
-            .then((response) => {
-                setFliprs(response.data)
-            })
-            .catch(e => console.error(e))
-    }
+    const {fliprs} = useFliprs()
 
     return(
         <FliprGallery fliprs={fliprs} />
